move test_parse_toa_line_exceptions to test_toa_reader.py
